requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-money==1.3.1
django-moneyfield==0.2.1
django-redis==4.12.1
django-versatileimagefield==2.0
djangorestframework==3.12.2
drf-jwt==1.18.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
Pillow==8.1.2
py-moneyed==0.8.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
python-magic==0.4.22
python-magic-bin==0.4.14
pytz==2021.1
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
